tutorial and cart issue fixed

+++ lib/network/graph_ql_manager.dart
@@ -137,7 +137,7 @@ class GraphQLManager {
 
   Future<QueryResult> createOrder(
     String userId,
-    double totalAmount,
+    int totalAmount,
     String deliveryType,
     String discountCode,
     String deliveryInstructions,
